project

+++ Coupons/src/main/java/coupons/logic/CouponController.java
@@ -54,7 +54,7 @@ public class CouponController {
 	public long createCoupon(Coupon coupon) throws ApplicationException {
 
 		if (coupon == null)
-			throw new ApplicationException(ErrorType.EMPTY.getMessage());
+			throw new ApplicationException(ErrorType.EMPTY, ErrorType.EMPTY.getMessage(), false);
 
 		ValidationUtils.isValidId(coupon.getCompanyId());
 		isValidCategory(coupon.getCategory());
@@ -65,25 +65,37 @@ public class CouponController {
 		isValidImage(coupon.getImage());
 
 		if (couponsDao.isCouponExists(coupon.getId()))
-			throw new ApplicationException(ErrorType.COUPON_IS_ALREADY_EXISTS.getMessage());
+			throw new ApplicationException(ErrorType.COUPON_IS_ALREADY_EXISTS,
+					ErrorType.COUPON_IS_ALREADY_EXISTS.getMessage(), false);
 
 		if (couponsDao.isCouponExists(coupon.getCompanyId(), coupon.getTitle()))
-			throw new ApplicationException(ErrorType.COUPON_IS_ALREADY_EXISTS.getMessage());
+			throw new ApplicationException(ErrorType.COUPON_IS_ALREADY_EXISTS,
+					ErrorType.COUPON_IS_ALREADY_EXISTS.getMessage(), false);
 
 		return couponsDao.createCoupon(coupon);
 
 	}
 
 	/**
-	 * @param couponId Receive a coupon id
+	 * @param couponId  Receive a coupon id
+	 * @param companyId Receive a company id
 	 * @throws ApplicationException This function can throw an applicationException
 	 */
-	public void deleteCoupon(long couponId) throws ApplicationException {
+	public void deleteCoupon(long couponId, long companyId) throws ApplicationException {
 
 		ValidationUtils.isValidId(couponId);
+		ValidationUtils.isValidId(companyId);
 
 		if (!couponsDao.isCouponExists(couponId))
-			throw new ApplicationException(ErrorType.COUPON_IS_NOT_EXISTS.getMessage());
+			throw new ApplicationException(ErrorType.COUPON_IS_NOT_EXISTS, ErrorType.COUPON_IS_NOT_EXISTS.getMessage(),
+					false);
+
+		if (!companyDao.isCompanyExists(companyId))
+			throw new ApplicationException(ErrorType.COMPANY_IS_NOT_EXISTS,
+					ErrorType.COMPANY_IS_NOT_EXISTS.getMessage(), false);
+
+		if (!couponsDao.isCouponBelongToCompany(couponId, companyId))
+			throw new ApplicationException(ErrorType.GENERAL_ERROR, ErrorType.GENERAL_ERROR.getMessage(), true);
 
 		purchasesDao.deletePurchaseByCouponId(couponId);
 
@@ -109,7 +121,7 @@ public class CouponController {
 	public void updateCoupon(Coupon coupon) throws ApplicationException {
 
 		if (coupon == null)
-			throw new ApplicationException(ErrorType.EMPTY.getMessage());
+			throw new ApplicationException(ErrorType.EMPTY, ErrorType.EMPTY.getMessage(), false);
 
 		ValidationUtils.isValidId(coupon.getId());
 		ValidationUtils.isValidId(coupon.getCompanyId());
@@ -121,7 +133,8 @@ public class CouponController {
 		isValidImage(coupon.getImage());
 
 		if (!couponsDao.isCouponExists(coupon.getId()))
-			throw new ApplicationException(ErrorType.COUPON_IS_NOT_EXISTS.getMessage());
+			throw new ApplicationException(ErrorType.COUPON_IS_NOT_EXISTS, ErrorType.COUPON_IS_NOT_EXISTS.getMessage(),
+					false);
 
 		Coupon couponFromDataBase = couponsDao.getCoupon(coupon.getId());
 
@@ -129,7 +142,8 @@ public class CouponController {
 
 			if (couponsDao.isCouponExists(coupon.getCompanyId(), coupon.getTitle()))
 
-				throw new ApplicationException(ErrorType.COUPON_IS_ALREADY_EXISTS.getMessage());
+				throw new ApplicationException(ErrorType.COUPON_IS_ALREADY_EXISTS,
+						ErrorType.COUPON_IS_ALREADY_EXISTS.getMessage(), false);
 
 		}
 
@@ -159,7 +173,8 @@ public class CouponController {
 		if (couponsDao.isCouponExists(couponId))
 			return couponsDao.getCoupon(couponId);
 
-		throw new ApplicationException(ErrorType.COUPON_IS_NOT_EXISTS.getMessage());
+		throw new ApplicationException(ErrorType.COUPON_IS_NOT_EXISTS, ErrorType.COUPON_IS_NOT_EXISTS.getMessage(),
+				false);
 	}
 
 	/**
@@ -172,7 +187,8 @@ public class CouponController {
 		ValidationUtils.isValidId(couponId);
 
 		if (!couponsDao.isCouponExists(couponId))
-			throw new ApplicationException(ErrorType.COUPON_IS_NOT_EXISTS.getMessage());
+			throw new ApplicationException(ErrorType.COUPON_IS_NOT_EXISTS, ErrorType.COUPON_IS_NOT_EXISTS.getMessage(),
+					false);
 
 		return couponsDao.howMuchCouponRemain(couponId);
 
@@ -188,7 +204,8 @@ public class CouponController {
 		ValidationUtils.isValidId(companyId);
 
 		if (!companyDao.isCompanyExists(companyId))
-			throw new ApplicationException(ErrorType.COMPANY_IS_NOT_EXISTS.getMessage());
+			throw new ApplicationException(ErrorType.COMPANY_IS_NOT_EXISTS,
+					ErrorType.COMPANY_IS_NOT_EXISTS.getMessage(), false);
 
 		return couponsDao.getCompanyCouponsByCompanyId(companyId);
 
@@ -206,7 +223,8 @@ public class CouponController {
 		isValidCategory(category);
 
 		if (!companyDao.isCompanyExists(companyId))
-			throw new ApplicationException(ErrorType.COMPANY_IS_NOT_EXISTS.getMessage());
+			throw new ApplicationException(ErrorType.COMPANY_IS_NOT_EXISTS,
+					ErrorType.COMPANY_IS_NOT_EXISTS.getMessage(), false);
 
 		return couponsDao.getCompanyCouponsByCategory(companyId, category);
 
@@ -224,7 +242,8 @@ public class CouponController {
 		isValidPrice(maxPrice);
 
 		if (!companyDao.isCompanyExists(companyId))
-			throw new ApplicationException(ErrorType.COMPANY_IS_NOT_EXISTS.getMessage());
+			throw new ApplicationException(ErrorType.COMPANY_IS_NOT_EXISTS,
+					ErrorType.COMPANY_IS_NOT_EXISTS.getMessage(), false);
 
 		return couponsDao.getCompanyCouponsByMaxPrice(companyId, maxPrice);
 
@@ -240,7 +259,8 @@ public class CouponController {
 		ValidationUtils.isValidId(customerId);
 
 		if (!customerDao.isCustomerExists(customerId))
-			throw new ApplicationException(ErrorType.CUSTOMER_IS_NOT_EXISTS.getMessage());
+			throw new ApplicationException(ErrorType.CUSTOMER_IS_NOT_EXISTS,
+					ErrorType.CUSTOMER_IS_NOT_EXISTS.getMessage(), false);
 
 		return couponsDao.getCustomerCouponsByCustomerId(customerId);
 
@@ -258,7 +278,8 @@ public class CouponController {
 		isValidCategory(category);
 
 		if (!customerDao.isCustomerExists(customerId))
-			throw new ApplicationException(ErrorType.CUSTOMER_IS_NOT_EXISTS.getMessage());
+			throw new ApplicationException(ErrorType.CUSTOMER_IS_NOT_EXISTS,
+					ErrorType.CUSTOMER_IS_NOT_EXISTS.getMessage(), false);
 
 		return couponsDao.getCustomerCouponsByCategory(customerId, category);
 
@@ -276,7 +297,8 @@ public class CouponController {
 		isValidPrice(maxPrice);
 
 		if (!customerDao.isCustomerExists(customerId))
-			throw new ApplicationException(ErrorType.CUSTOMER_IS_NOT_EXISTS.getMessage());
+			throw new ApplicationException(ErrorType.CUSTOMER_IS_NOT_EXISTS,
+					ErrorType.CUSTOMER_IS_NOT_EXISTS.getMessage(), false);
 
 		return couponsDao.getCustomerCouponsByMaxPrice(customerId, maxPrice);
 
@@ -289,7 +311,7 @@ public class CouponController {
 	private void isValidPrice(double price) throws ApplicationException {
 
 		if (price <= 0)
-			throw new ApplicationException(ErrorType.INVALID_PRICE.getMessage());
+			throw new ApplicationException(ErrorType.INVALID_PRICE, ErrorType.INVALID_PRICE.getMessage(), false);
 
 	}
 
@@ -300,7 +322,7 @@ public class CouponController {
 	private void isValidCategory(Category category) throws ApplicationException {
 
 		if (category == null)
-			throw new ApplicationException(ErrorType.INVALID_CATEGORY.getMessage());
+			throw new ApplicationException(ErrorType.INVALID_CATEGORY, ErrorType.INVALID_CATEGORY.getMessage(), false);
 
 	}
 
@@ -311,7 +333,7 @@ public class CouponController {
 	private void isValidImage(String image) throws ApplicationException {
 
 		if (!(image.contains(".")) || image.charAt(image.length() - 1) == '.' || image.charAt(0) == '.')
-			throw new ApplicationException(ErrorType.INVALID_IMAGE.getMessage());
+			throw new ApplicationException(ErrorType.INVALID_IMAGE, ErrorType.INVALID_IMAGE.getMessage(), false);
 
 	}
 }
